Refactor/tests (#8679)
* moving artifactory

* minor refactors

* review, moving tools.vcvars to another folder
