+++ conans/test/functional/tools/old/vcvars/vcvars_store_test.py
